Rewrite patricia tree merge functions with `pattern_match_pair` (#4262)
This is a rewrite of both #4073 and #4183.

Instead of writing a `universal_merge` function in the style of Haskell's `Data.IntMap`, we use a `pattern_match_pair` function encapsulating the logic for pattern-matching on pairs of trees. This is both more flexible, and avoids leaving unneeded value slots around after simplification as with the `universal_merge` function.
 
Each of the function (`union`, etc.) is itself recursive, but the code remains small and readable:

```ocaml
  let rec union f t0 t1 =
    let iv = is_value_of t0 in
    pattern_match_pair_merge
      ~only_left:(fun t0 -> t0)
      ~only_right:(fun t1 -> t1)
      ~both_sides:(fun t0 t1 -> union f t0 t1)
      iv f t0 t1
```

For `union` specifically, I have carefully made sure that the generated code is identical to the current code in `main` (in fact, a tiny bit better, because we are able to avoid one allocation we failed to eliminate before).

This function includes both the `diff` function from #4073 (which is used in #4112) and the `update_many` function from #4183 (which will be used in a separate PR to optimize the quadratic use case in dataflow mentioned in #4183).
---------

Co-authored-by: Pierre Chambart <pierre.chambart@ocamlpro.com>