Add support for the `[@atomic]` attribute on mutable fields (#4132)
Add support to only the frontend for specifying the `[@atomic]` attribute on
(only) mutable fields. This is a no-op currently, but is forwarded down through
the ld_mutable on the label description in the typedtree.

This is the first of many commits pulling in upstream support for atomic
record fields. I've slightly diverged from the design there, by adding
atomic_flag to mutability only under Mutable rather than as a new ld_atomic
field, since that feels like a better design (it avoids immutable atomic fields
by construction, rather than having to assert that they never happen)


---------

Co-authored-by: Richard Eisenberg <reisenberg@janestreet.com>