Fix soundness issue with recursive abstract with bounds (#4435)
* Fix soundness issue with recursive abstract with bounds found by Benjamin Peters.

Example of the bug:
    module type X = sig
      type t : value mod contended with t
    end
    module Xm : X = struct
      type t = int ref
    end
    type q : value mod contended = Xm.t  (* unsoundly accepted *)

The cause of the bug is that the mechanism to compute recursive jkinds optimistically interacted poorly with abstract types. For recursive concrete kinds, we want to pick the best kind we can. For recursive abstract kinds, we need to pessimistically assume that the implementor of the signature chooses the worst possible kind. The previous code chose optimistically in both cases.

This commit fixes the issue and rejects the code above as kind-incorrect.