Dedicated engine for Cmm Peephole optimisations (#4118)
This module was initially merged with support for a subset of the Cmm
peephole optimisations performed in Cmm_helpers. It was meant to be extended
incrementally, and the individual commits adding support for the various
operations were meant to serve as reference for further extensions. These
individual commits have disappeared from the main history, but they have been
preserved in the cmm-peephole-engine branch of the main repo, accessible at
<https://github.com/oxcaml/oxcaml/tree/cmm-peephole-engine>. 

Co-authored-by: Nathanaëlle Courant <Ekdohibs@users.noreply.github.com>
Co-authored-by: Guillaume Bury <guillaume.bury@gmail.com>