Add support for a new [%atomic.loc r.field] construct (#4229)
* Add atomic_loc to predef

* Add support for a new [%atomic.loc r.field] construct

Add support for a new [%atomic.lock r.field] contstruct, which gets typed into a
new Texp_atomic_loc typedtree expression node, which in turn turns into a block
allocation containing the record `r`, and the field index of the field `field`
in that record, with type `'a atomic_loc`, where `'a'` is the type of that
field:

Γ ⊢ e₁ : τ₁ @ m,
     τ₁ = { ... f : τ₂ ... }
——————————————————————————————————————————– ATOMIC_LOC
Γ ⊢ [%atomic.loc e₁.f] : τ₂ atomic_loc @ m

* Don't apply modalities to atomic_loc

We now forbid anything besides legacy, so we can just ignore it here and use the
mode of the record

* Don't mode cross for the record

Otherwise if the field mode crosses locality we'd get a global loc with a
pointer to a local record, which is no good.

* Forbid non-legacy modalities on fields given to atomic.loc

If we allow changing the modalities here, we'd need a way to restrict the atomic
loc primitives to changing the fields to values at compatible modes, which we
don't currently have a nice way to do. Maybe, someday, first-class modalities
will save us, but in the meantime the safest option here is to just forbid this.

A past version of this commit made this change at type declaration site, but it
was pointed out that it's generally better to restrict only the operation we
want.


---------

Co-authored-by: Zesen Qian <github@riaqn.org>